+++ cards/ephemeral-realm/writer.ts
@@ -2,7 +2,7 @@ import { Writer } from '@cardstack/core/writer';
 import { Session } from '@cardstack/core/session';
 import { UpstreamDocument, UpstreamIdentity } from '@cardstack/core/document';
 import { inject } from '@cardstack/hub/dependency-injection';
-import { AddressableCard } from '@cardstack/core/card';
+import { AddressableCard } from '@cardstack/hub';
 import CardstackError from '@cardstack/core/error';
 
 let counter = 0;
