+++ cards/git-realm/indexer.ts
@@ -8,8 +8,8 @@ import service from './lib/service';
 
 const defaultBranch = 'master';
 
-import { Indexer, IndexingOperations } from '@cardstack/core/indexer';
-import { AddressableCard, Card } from '@cardstack/core/card';
+import { Indexer, IndexingOperations } from '@cardstack/hub';
+import { AddressableCard, Card } from '@cardstack/hub';
 import { extractSettings } from './lib/git-settings';
 import { assertSingleResourceDoc } from '@cardstack/core/jsonapi';
 import { SingleResourceDoc } from 'jsonapi-typescript';
