+++ packages/cardhost/app/routes/cards/card/card-model-route.ts
@@ -1,7 +1,7 @@
 import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
 import { CardstackSession } from '../../../services/cardstack-session';
-import { Card } from '@cardstack/core/card';
+import { Card } from '@cardstack/hub';
 
 export default class CardModelRoute extends Route {
   @service cardstackSession!: CardstackSession;
