+++ packages/cardhost/app/services/css.ts
@@ -1,6 +1,6 @@
 import Service from '@ember/service';
 import { tracked } from '@glimmer/tracking';
-import { AddressableCard } from '@cardstack/core/card';
+import { AddressableCard } from '@cardstack/hub';
 import sortBy from 'lodash/sortBy';
 import { set } from '@ember/object';
