+++ packages/core/module-loader.ts
@@ -1,4 +1,4 @@
-import { Card } from './card';
+import { Card } from '@cardstack/hub';
 
 export interface ModuleLoader {
   load(card: Card, localModulePath: string, exportedName?: string): Promise<any>;
