+++ packages/hub/cards-service.ts
@@ -1,5 +1,5 @@
 import { Session } from '@cardstack/core/session';
-import { UnsavedCard, AddressableCard, Card } from '@cardstack/core/card';
+import { UnsavedCard, AddressableCard, Card } from './card';
 import { asCardId, CardId } from '@cardstack/core/card-id';
 import { CARDSTACK_PUBLIC_REALM } from '@cardstack/core/realm';
 import { ResponseMeta } from '@cardstack/core/document';
